Rename dataset and model files to conform to module formatting
